Customization layout has been removed. Code structure has been refined
